broke sensor_control up into different files
